implement auto-split weight calculation, refactoring for consistency

+++ ganilib/Classes/Screen/GScreen.swift
@@ -101,7 +101,7 @@ open class GScreen: UIViewController {
         return self
     }
     
-    public func paddings(t top: CGFloat? = nil, l left: CGFloat? = nil, b bottom: CGFloat? = nil, r right: CGFloat? = nil) -> Self {
+    public func paddings(t top: Float? = nil, l left: Float? = nil, b bottom: Float? = nil, r right: Float? = nil) -> Self {
         if let view = screenContent() as? GContainer {
             _ = view.paddings(t: top, l: left, b: bottom, r: right)
         }
