+++ ganilib/Classes/View/Layout/GScrollView.swift
@@ -66,7 +66,7 @@ open class GScrollView: UIScrollView {
         contentView.addView(view, top: top)
     }
     
-    public func paddings(t top: CGFloat? = nil, l left: CGFloat? = nil, b bottom: CGFloat? = nil, r right: CGFloat? = nil) -> Self {
+    public func paddings(t top: Float? = nil, l left: Float? = nil, b bottom: Float? = nil, r right: Float? = nil) -> Self {
         _ = contentView.paddings(t: top, l: left, b: bottom, r: right)
         return self
     }
