+++ ganilib/Classes/View/Table/GTableView.swift
@@ -89,7 +89,7 @@ open class GTableView: UITableView, GContainer {
         return type.init(style: style)
     }
     
-    public func paddings(t top: CGFloat? = nil, l left: CGFloat? = nil, b bottom: CGFloat? = nil, r right: CGFloat? = nil) -> Self {
+    public func paddings(t top: Float? = nil, l left: Float? = nil, b bottom: Float? = nil, r right: Float? = nil) -> Self {
         helper.paddings(t: top, l: left, b: bottom, r: right)
         return self
     }
@@ -123,7 +123,7 @@ open class GTableViewCell: UITableViewCell {
 //        self.contentView.addSubview(container)
     }
     
-    public func paddings(t top: CGFloat? = nil, l left: CGFloat? = nil, b bottom: CGFloat? = nil, r right: CGFloat? = nil) -> Self {
+    public func paddings(t top: Float? = nil, l left: Float? = nil, b bottom: Float? = nil, r right: Float? = nil) -> Self {
         helper.paddings(t: top, l: left, b: bottom, r: right)
         return self
     }
@@ -252,7 +252,7 @@ open class GHeaderFooterView: UIView {
         return self
     }
     
-    public func paddings(t top: CGFloat? = nil, l left: CGFloat? = nil, b bottom: CGFloat? = nil, r right: CGFloat? = nil) -> Self {
+    public func paddings(t top: Float? = nil, l left: Float? = nil, b bottom: Float? = nil, r right: Float? = nil) -> Self {
         _ = container.paddings(t: top, l: left, b: bottom, r: right)
         return self
     }
